renamed debug/sys to s_fps
