+++ voxels/src/shared/debug/mod.rs
@@ -1,3 +1,3 @@
-mod sys;
+mod s_fps;
 
-pub use sys::*;
+pub use s_fps::*;
